+++ go.mod
@@ -4,6 +4,7 @@ go 1.14
 
 require (
 	github.com/dgrijalva/jwt-go v3.2.0+incompatible
+	github.com/go-chi/chi v4.1.2+incompatible
 	github.com/google/uuid v1.1.1
 	github.com/mattn/go-sqlite3 v1.14.0
 )
