+++ internal/transport/togo_handler.go
@@ -1,10 +1,224 @@
 package transport
 
 import (
-	"github.com/manabie-com/togo/internal/usecase"
+	"context"
+	"database/sql"
+	"encoding/json"
+	"log"
+	"net/http"
+	"time"
+
+	"github.com/dgrijalva/jwt-go"
+	"github.com/go-chi/chi"
+	"github.com/go-chi/chi/middleware"
+	"github.com/google/uuid"
+	togo "github.com/manabie-com/togo/internal"
+	"github.com/manabie-com/togo/internal/entities"
+	"github.com/manabie-com/togo/internal/storages"
 )
 
 //TogoHandler represent the httphandler for togo
 type TogoHandler struct {
-	TogoUsecase usecase.TogoUsecase
+	togoUsecase togo.Usecase
+	JWTKey      string
+}
+
+func NewTogoHandler(mux *chi.Mux, us togo.Usecase, JWTKey string) {
+	handler := &TogoHandler{
+		togoUsecase: us,
+		JWTKey:      JWTKey,
+	}
+	// StripSlashes remove redundant slash in endpoint, example /login/ -> /login
+	mux.Use(middleware.StripSlashes)
+
+	mux.Get("/login", handler.getAuthToken)
+	mux.Group(func(r chi.Router) {
+		r.Use(handler.authMiddleware)
+		r.Get("/tasks", handler.listTasks)
+		r.Post("/tasks", handler.addTask)
+
+	})
+}
+func (t *TogoHandler) authMiddleware(next http.Handler) http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		var ok bool
+		r, ok = t.validToken(r)
+		if !ok {
+			w.WriteHeader(http.StatusUnauthorized)
+			return
+		}
+		next.ServeHTTP(w, r)
+	})
+}
+
+type ToDoService struct {
+	JWTKey string
+	DB     storages.Storages
+}
+
+// func (s *ToDoService) ServeHTTP(resp http.ResponseWriter, req *http.Request) {
+// 	log.Println(req.Method, req.URL.Path)
+// 	resp.Header().Set("Access-Control-Allow-Origin", "*")
+// 	resp.Header().Set("Access-Control-Allow-Headers", "*")
+// 	resp.Header().Set("Access-Control-Allow-Methods", "*")
+
+// 	if req.Method == http.MethodOptions {
+// 		resp.WriteHeader(http.StatusOK)
+// 		return
+// 	}
+
+// 	switch req.URL.Path {
+// 	case "/login":
+// 		s.getAuthToken(resp, req)
+// 		return
+// 	case "/tasks":
+// 		var ok bool
+// 		req, ok = s.validToken(req)
+// 		if !ok {
+// 			resp.WriteHeader(http.StatusUnauthorized)
+// 			return
+// 		}
+
+// 		switch req.Method {
+// 		case http.MethodGet:
+// 			s.listTasks(resp, req)
+// 		case http.MethodPost:
+// 			s.addTask(resp, req)
+// 		}
+// 		return
+// 	}
+// }
+
+func (t *TogoHandler) getAuthToken(resp http.ResponseWriter, req *http.Request) {
+	id := value(req, "user_id")
+	if !t.togoUsecase.ValidateUser(req.Context(), id, value(req, "password")) {
+		resp.WriteHeader(http.StatusUnauthorized)
+		json.NewEncoder(resp).Encode(map[string]string{
+			"error": "incorrect user_id/pwd",
+		})
+		return
+	}
+	resp.Header().Set("Content-Type", "application/json")
+
+	token, err := t.createToken(id.String)
+	if err != nil {
+		resp.WriteHeader(http.StatusInternalServerError)
+		json.NewEncoder(resp).Encode(map[string]string{
+			"error": err.Error(),
+		})
+		return
+	}
+
+	json.NewEncoder(resp).Encode(map[string]string{
+		"data": token,
+	})
+}
+
+func (t *TogoHandler) listTasks(resp http.ResponseWriter, req *http.Request) {
+	id, _ := userIDFromCtx(req.Context())
+	tasks, err := t.togoUsecase.RetrieveTasks(
+		req.Context(),
+		sql.NullString{
+			String: id,
+			Valid:  true,
+		},
+		value(req, "created_date"),
+	)
+
+	resp.Header().Set("Content-Type", "application/json")
+
+	if err != nil {
+		resp.WriteHeader(http.StatusInternalServerError)
+		json.NewEncoder(resp).Encode(map[string]string{
+			"error": err.Error(),
+		})
+		return
+	}
+
+	json.NewEncoder(resp).Encode(map[string][]*entities.Task{
+		"data": tasks,
+	})
+}
+
+func (t *TogoHandler) addTask(resp http.ResponseWriter, req *http.Request) {
+	task := &entities.Task{}
+	err := json.NewDecoder(req.Body).Decode(t)
+	defer req.Body.Close()
+	if err != nil {
+		resp.WriteHeader(http.StatusInternalServerError)
+		return
+	}
+
+	now := time.Now()
+	userID, _ := userIDFromCtx(req.Context())
+	task.ID = uuid.New().String()
+	task.UserID = userID
+	task.CreatedDate = now.Format("2006-01-02")
+
+	resp.Header().Set("Content-Type", "application/json")
+
+	err = t.togoUsecase.AddTask(req.Context(), task)
+	if err != nil {
+		resp.WriteHeader(http.StatusInternalServerError)
+		json.NewEncoder(resp).Encode(map[string]string{
+			"error": err.Error(),
+		})
+		return
+	}
+
+	json.NewEncoder(resp).Encode(map[string]*entities.Task{
+		"data": task,
+	})
+}
+
+func value(req *http.Request, p string) sql.NullString {
+	return sql.NullString{
+		String: req.FormValue(p),
+		Valid:  true,
+	}
+}
+
+func (t *TogoHandler) createToken(id string) (string, error) {
+	atClaims := jwt.MapClaims{}
+	atClaims["user_id"] = id
+	atClaims["exp"] = time.Now().Add(time.Minute * 15).Unix()
+	at := jwt.NewWithClaims(jwt.SigningMethodHS256, atClaims)
+	token, err := at.SignedString([]byte(t.JWTKey))
+	if err != nil {
+		return "", err
+	}
+	return token, nil
+}
+
+func (s *TogoHandler) validToken(req *http.Request) (*http.Request, bool) {
+	token := req.Header.Get("Authorization")
+
+	claims := make(jwt.MapClaims)
+	t, err := jwt.ParseWithClaims(token, claims, func(*jwt.Token) (interface{}, error) {
+		return []byte(s.JWTKey), nil
+	})
+	if err != nil {
+		log.Println(err)
+		return req, false
+	}
+
+	if !t.Valid {
+		return req, false
+	}
+
+	id, ok := claims["user_id"].(string)
+	if !ok {
+		return req, false
+	}
+
+	req = req.WithContext(context.WithValue(req.Context(), userAuthKey(0), id))
+	return req, true
+}
+
+type userAuthKey int8
+
+func userIDFromCtx(ctx context.Context) (string, bool) {
+	v := ctx.Value(userAuthKey(0))
+	id, ok := v.(string)
+	return id, ok
 }
