+++ internal/usecase/togo_ucase.go
@@ -8,16 +8,20 @@ import (
 	"github.com/manabie-com/togo/internal/storages"
 )
 
-type TogoUsecase interface {
-	RetrieveTasks(ctx context.Context, userID, createdDate sql.NullString) ([]*entities.Task, error)
-	AddTask(ctx context.Context, t *entities.Task) error
-	ValidateUser(ctx context.Context, userID, pwd sql.NullString) bool
-}
-
 type togoUsecase struct {
-	togoStorages storages.DB
+	togoStorages storages.Storages
 }
 
-func NewTogoUsecase(d storages.DB) TogoUsecase {
-
+// NewTogoUsecase will create new a togoUsecase object representation of togo.Usecase interface
+func NewTogoUsecase(d storages.Storages) togoUsecase {
+	return togoUsecase{togoStorages: d}
+}
+func (t *togoUsecase) RetrieveTasks(ctx context.Context, userID, createdDate sql.NullString) ([]*entities.Task, error) {
+	return t.togoStorages.RetrieveTasks(context.Background(), userID, createdDate)
+}
+func (t *togoUsecase) AddTask(ctx context.Context, task *entities.Task) error {
+	return t.togoStorages.AddTask(ctx, task)
+}
+func (t *togoUsecase) ValidateUser(ctx context.Context, userID, pwd sql.NullString) bool {
+	return t.togoStorages.ValidateUser(ctx, userID, pwd)
 }
